adding pass view

+++ api/api.go
@@ -2,10 +2,10 @@ package api
 
 type Station struct {
 	StnName    string  `json:"stnname,omitempty" db:"stnname"`
-	Latitude   float32 `json:"latitude,omitempty" db:"latitude"`
-	Longitude  float32 `json:"longitude,omitempty" db:"longitude"`
-	Altitude   float32 `json:"altitude,omitempty" db:"altitude"`
-	MinHorizon float32 `json:"minhorizon,omitempty" db:"minhorizon"`
+	Latitude   float64 `json:"latitude,omitempty" db:"latitude"`
+	Longitude  float64 `json:"longitude,omitempty" db:"longitude"`
+	Altitude   float64 `json:"altitude,omitempty" db:"altitude"`
+	MinHorizon float64 `json:"minhorizon,omitempty" db:"minhorizon"`
 }
 
 type TLE struct {
